+++ app/controllers/furimas_controller.rb
@@ -1,27 +1,2 @@
 class FurimasController < ApplicationController
-  before_action :authenticate_user!, except: [:index, :show]
-
-  def index
-    @items = Item.all.order(created_at: :desc)
-  end
-
-  def new
-    @item = Item.new
-  end
-
-  def create
-    @item = current_user.items.new(item_params)
-    if @item.save
-      redirect_to root_path
-    else
-      render :new, status: :unprocessable_entity
-    end
-  end  
-
-  private
-
-  def item_params
-    params.require(:item).permit(:name, :description, :category_id, :condition_id, :shipping_fee_id, :prefecture_id, :shipping_day_id, :price, :image).merge(user_id: current_user.id)
-  end
-
 end
