+++ app/helpers/items_helper.rb
@@ -1,0 +1,2 @@
+module ItemsHelper
+end
