+++ app/views/items/index.html.erb
@@ -129,7 +129,7 @@
   <ul class='item-lists'>
 <% @items.each do |item| %>
   <li class='list'>
-    <%# <%=# link_to item_path(item.id) do %>
+      <%= link_to item_path(item.id) do %>
       <div class='item-img-content'>
         <%= image_tag item.image, class: "item-img" %>
 
@@ -149,7 +149,7 @@
             </div>
           </div>
         </div>
-      <%#<% end %>
+      <% end %>
     </div>
   </li>
 <% end %>
@@ -176,7 +176,7 @@
 <% end %>
 </ul>
 </div>
-<%= link_to new_furima_path, class: 'purchase-btn' do %>
+<%= link_to new_item_path, class: 'purchase-btn' do %>
   <span class='purchase-btn-text'>出品する</span>
   <%= image_tag 'icon_camera.png' , size: '185x50' ,class: "purchase-btn-icon" %>
 <% end %>
