+++ spec/requests/items_request_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe "Items", type: :request do
+
+end
